feat: create basic docker setup
But uh hey, it mostly works?

WARNING:
* Events duplicate and continue to exist and duplicate after each docker-compose up and
shutdown. You can clear it through running docker-compose down or
pretending it doesn't exist.

NOTES:
* The stuff with the javascript mongo-init script makes it work
somehow, I'm not really sure.
* I made it so that the DB would not be loaded when actually building
the site so that I can run npm run build and serve it through npm run
preview. I'm not sure if this breaks vercel.
* I also changed the db connection thing to use DB_NAME once instead
of twice because that's just confusing.
* There's also a lot of new methods under /db/db.ts to populate the
parts of it.